Move Hubspot functions to helpers, getCookie function added.
